Modified SSTO problem to be a single component rather than a group
