+++ dymos/examples/ssto/doc/test_doc_ssto_earth.py
@@ -1,8 +1,6 @@
 import unittest
 
 import matplotlib.pyplot as plt
-plt.switch_backend('Agg')
-plt.style.use('ggplot')
 
 from openmdao.utils.assert_utils import assert_near_equal
 from openmdao.utils.testing_utils import use_tempdirs
@@ -23,7 +21,7 @@ class TestDocSSTOEarth(unittest.TestCase):
         #
         p = om.Problem(model=om.Group())
         p.driver = om.pyOptSparseDriver()
-        p.driver.declare_coloring()
+        p.driver.declare_coloring(tol=1.0E-12)
 
         from dymos.examples.ssto.launch_vehicle_ode import LaunchVehicleODE
 
@@ -45,18 +43,18 @@ class TestDocSSTOEarth(unittest.TestCase):
         phase.set_time_options(initial_bounds=(0, 0), duration_bounds=(10, 500))
 
         phase.add_state('x', fix_initial=True, ref=1.0E5, defect_ref=1.0,
-                        rate_source='eom.xdot', units='m')
+                        rate_source='xdot', units='m')
         phase.add_state('y', fix_initial=True, ref=1.0E5, defect_ref=1.0,
-                        rate_source='eom.ydot', targets=['atmos.y'], units='m')
+                        rate_source='ydot', targets=['y'], units='m')
         phase.add_state('vx', fix_initial=True, ref=1.0E3, defect_ref=1.0,
-                        rate_source='eom.vxdot', targets=['eom.vx'], units='m/s')
+                        rate_source='vxdot', targets=['vx'], units='m/s')
         phase.add_state('vy', fix_initial=True, ref=1.0E3, defect_ref=1.0,
-                        rate_source='eom.vydot', targets=['eom.vy'], units='m/s')
+                        rate_source='vydot', targets=['vy'], units='m/s')
         phase.add_state('m', fix_initial=True, ref=1.0E3, defect_ref=1.0,
-                        rate_source='eom.mdot', targets=['eom.m'], units='kg')
+                        rate_source='mdot', targets=['m'], units='kg')
 
-        phase.add_control('theta', units='rad', lower=-1.57, upper=1.57, targets=['eom.theta'])
-        phase.add_parameter('thrust', units='N', opt=False, val=2100000.0, targets=['eom.thrust'])
+        phase.add_control('theta', units='rad', lower=-1.57, upper=1.57, targets=['theta'])
+        phase.add_parameter('thrust', units='N', opt=False, val=2100000.0, targets=['thrust'])
 
         #
         # Set the options for our constraints and objective
